refactor: move all abstract conversion logic under conversion
